Added the powerline role
This role is meant to be a flexible way to install ansible in a system.
Currently the role is being prototyped for adding powerline to shell
terminals (e.g bash, zsh, etc..) on Ubuntu-based distros.

The intention is to expand it to also include the setup VIM to use
powerline.

Also, added some group_vars configuration for the control group (the
control node) through the control.yml file.